+++ Assets/Mu3LibraryAssets/Runtime/Scripts/Observable/IObservableValue.cs
@@ -1,0 +1,9 @@
+namespace Mu3Library.Observable
+{
+    public interface IObservableValue<out TValue>
+    {
+        TValue Value { get; }
+
+        System.IDisposable Subscribe(System.Action<TValue> callback, bool notifyImmediately = false);
+    }
+}